+++ Packages/com.stackray.text/Stackray.Text/VertexIndex.cs
@@ -3,8 +3,8 @@
 namespace Stackray.Text {
   [InternalBufferCapacity(54)]    // 54 is the worst case scenario for 9-slice sprite
   public struct VertexIndex : IBufferElementData {
-    public ushort Value;
-    public static implicit operator VertexIndex(int v) { return new VertexIndex { Value = (ushort)v }; }
+    public int Value;
+    public static implicit operator VertexIndex(int v) { return new VertexIndex { Value = v }; }
     public static implicit operator int(VertexIndex v) { return v.Value; }
   }
 }
